requirements.txt
django==4.2.7
python-dotenv==1.0.0
pre-commit==3.5.0
djangorestframework==3.14.0
